use a generalized solrizable_resource
